Calculadora_Somar_RetornarValoresCorretos

+++ TesteDeSoftware/Demo/Calculadora.cs
@@ -8,7 +8,7 @@ namespace Demo
 {
     public class Calculadora
     {
-        public double Somar(int v1, int v2) => v1 + v2;
+        public double Somar(double v1, double v2) => v1 + v2;
         public int Dividir(int v1, int v2)=>v2 / v1;
     }
 }
